requirements.txt
sentence-transformers
python-docx
numpy
pandas
